requirements.txt
slackclient
flask
slackeventsapi
python-dotenv